Fixing type

+++ MiniTabBar/Classes/MiniTabBar.swift
@@ -72,7 +72,7 @@ import UIKit
             }
         }
     }
-    private var positionY: Int
+    private var positionY: CGFloat
     private var animatedHide: Bool
     private let visualEffectView = UIVisualEffectView(effect: UIBlurEffect(style: .extraLight)) as UIVisualEffectView
     public var backgroundBlurEnabled: Bool = true {
